Fix exception that was causing the lag with cluster finding.

+++ src/com/colorbot/util/Point.java
@@ -2,17 +2,17 @@ package com.colorbot.util;
 
 public class Point {
 
-	public short x;
-	public short y;
+	public int x;
+	public int y;
 
 	public Point(int x, int y) {
-		this.x = (short) x;
-		this.y = (short) y;
+		this.x = x;
+		this.y = y;
 	}
 
 	public Point(java.awt.Point p) {
-		this.x = (short) p.x;
-		this.y = (short) p.y;
+		this.x = p.x;
+		this.y = p.y;
 	}
 	
 	@Override
